+++ cli/src/commands/git/init.rs
@@ -91,7 +91,7 @@ pub fn cmd_git_init(
     let cwd = command.cwd();
     let wc_path = cwd.join(&args.destination);
     let wc_path = file_util::create_or_reuse_dir(&wc_path)
-        .and_then(|_| wc_path.canonicalize())
+        .and_then(|_| dunce::canonicalize(wc_path))
         .map_err(|e| user_error_with_message("Failed to create workspace", e))?;
 
     do_init(