+++ CHANGELOG.md
@@ -81,6 +81,9 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 * `jj config path --user` no longer creates new file at the default config path.
 
+* On Windows, workspace paths (printed by `jj root`) no longer use UNC-style
+  `\\?\` paths unless necessary.
+
 ## [0.24.0] - 2024-12-04
 
 ### Release highlights